requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.0
fastapi==0.109.0
fasteners==0.14.1
future==0.18.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
systemd-python==234
tqdm==4.66.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.27.0
uvloop==0.19.0
vboxapi==1.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
xdg==5
xkit==0.0.0
zipp==1.0.0
